+++ web/app/countrytopos.js
@@ -324,7 +324,7 @@ exports.addCountryTopos = function(countries) {
     countries['QA'] = getCountry('QAT')
     countries['RE'] = getCountry('REU')
     countries['RO'] = getCountry('ROU')
-    // countries['RU'] = getCountry('RUS') ** BUGGY
+    countries['RU'] = getCountry('RUS')
     countries['RW'] = getCountry('RWA')
     countries['BL'] = getCountry('BLM')
     countries['SH'] = getCountry('SHN')
